Split mcp client into transport/protocol/builtins modules
